Move userslist to .api (was in .views)
