+++ app/minichat/api.py
@@ -1,13 +1,20 @@
-from .models import Message
 from rest_framework.serializers import ModelSerializer
 from rest_framework.viewsets import ReadOnlyModelViewSet
-from profile.api import UserSerializer
+from rest_framework.views import APIView
+from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.fields import CharField
+
+from .models import Message
+
+from profile.api import UserSerializer
+from profile.models import ActiveUser
+
 from minichat.templatetags.minichat import urlize3
 from commons.templatetags.markup_bbcode import smiley
-from rest_framework.fields import CharField
 from django.contrib.humanize.templatetags.humanize import naturalday
 from django.template.defaultfilters import time
+
 
 class MinichatTextField(CharField):
     def to_representation(self, value):
@@ -50,4 +57,25 @@ class LatestMessagesViewSet(ReadOnlyModelViewSet):
     queryset = Message.objects.order_by('-date')
     serializer_class = MessageSerializer
     pagination_class = LatestMessagesPagination
+
+class UsersListView(APIView):
+    """
+    Return a list of available users whose username starts with the value in `query`.
+    """
+
+    def get_queryset(self):
+        username = self.request.query_params.get('query', None)
+        if username and len(username) > 2:
+            qs = ActiveUser.objects.filter(username__istartswith=username[1:])
+        else:
+            qs = ActiveUser.objects.none()
+        return qs
+
+    def get(self, request, format=None):
+        qs = self.get_queryset()
+        usernames = ['@%s' % user.username for user in qs]
+        return Response({
+            'query': request.query_params.get('query', None),
+            'suggestions': usernames
+        })
